Fix circular import between EditableGrid.tsx and internal/actions.ts
