+++ packages/components/src/internal/util/utils.ts
@@ -20,6 +20,10 @@ import { hasParameter, toggleParameter } from '../..';
 
 const emptyList = List<string>();
 
+export function genCellKey(colIdx: number, rowIdx: number): string {
+    return [colIdx, rowIdx].join('-');
+}
+
 // Case insensitive Object reference. Returns undefined if either object or prop does not resolve.
 // If both casings exist (e.g. 'x' and 'X' are props) then either value may be returned.
 export function caseInsensitive(obj: Record<string, any>, prop: string): any {
